Move ingest fixtures to ndbfixtures
