+++ nucliadb/tests/ingest/integration/orm/test_orm_knowledgebox.py
@@ -31,7 +31,7 @@ from nucliadb_protos import utils_pb2
 from nucliadb_protos.knowledgebox_pb2 import SemanticModelMetadata
 from nucliadb_utils.storages.storage import Storage
 from nucliadb_utils.utilities import Utility, clean_utility, get_utility, set_utility
-from tests.ingest.fixtures import broker_resource
+from tests.ndbfixtures.ingest import broker_resource
 
 
 @pytest.fixture(scope="function")
