+++ nucliadb/tests/ingest/integration/orm/test_orm_resource.py
@@ -46,7 +46,7 @@ from nucliadb_protos.writer_pb2 import (
     FieldID,
     Paragraph,
 )
-from tests.ingest.fixtures import create_resource
+from tests.ndbfixtures.ingest import create_resource
 
 
 async def test_create_resource_orm_with_basic(storage, txn, cache, fake_node, knowledgebox_ingest: str):
